MFC r329067:
Fix "zpool add" crash when a replacing vdev has a spare child

Fix an assertion in zpool that causes a crash when running any "zpool add"
command on a spare that contains a replacing vdev with a spare child.

This likely affects Illumos, too.

PR:		225546
Sponsored by:	Spectra Logic Corp
Differential Revision:	https://reviews.freebsd.org/D14138


git-svn-id: https://svn.freebsd.org/base/stable/11@330728 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f