MFC r322277 by jlh:
rwho/ruptime/rwhod shouldn't be gated by RCMDS.

As peter@ points out in pr/220953:
"rwho, rwhod and ruptime are not part of the remote login suite (rsh, rlogin
etc).

They should *not* be in the rcmds package which is disabled by default.  We
rely on rwho/rwhod/ruptime in the freebsd.org cluster."

This commit is a re-commit of r322029 and r322031 with a better commit log, as
pointed out by ngie@.

This also includes the necesary changes to OptionalObsoleteFiles.inc, as
requested by jhb@.

PR:		220953


git-svn-id: https://svn.freebsd.org/base/stable/11@330769 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f