MFC r322278,324177: EFIRT Improvements
r322278 (imp): Fail to open efirt device when no EFI on system.

libefivar expects opening /dev/efi to indicate if the we can make efi
runtime calls. With a null routine, it was always succeeding leading
efi_variables_supported() to return the wrong value. Only succeed if
we have an efi_runtime table. Also, while I'm hear, out of an
abundance of caution, add a likely redundant check to make sure
efi_systbl is not NULL before dereferencing it. I know it can't be
NULL if efi_cfgtbl is non-NULL, but the compiler doesn't.

r324177 (andrew):
To prepare for adding EFI runtime services support on arm64 move the
machine independent parts of the existing code to a new file that can be
shared between amd64 and arm64.

Care has been taken to ensure that the MFC of r324177 did not clobber
cherry-picked MFC's.


git-svn-id: https://svn.freebsd.org/base/stable/11@331019 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f