MFC (partially) r326066, r326121: Add an EFI RTC Driver
This is effectively a direct commit to stable/11 to avoid the arm64 bits, as
those are unfortunately not able to MFC'd.

MFC r326066:
Add a driver for the EFI RTC. This uses the EFI Runtime Services to query
the system time.

MFC r326121:
Zero struct efi_tm before setting the needed values. We don't use the dst
or timezone fields so ensure these are set.


git-svn-id: https://svn.freebsd.org/base/stable/11@331025 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f