MFC r330843:
efirtc: Pass a dummy tmcap pointer to efi_get_time_locked

As noted in the comment, UEFI spec claims the capabilities pointer is
optional, but some implementations will choke and attempt to dereference it
without checking. This specific problem was found on a Lenovo Thinkpad X220
that would panic in efirtc_identify.

Requested by:	kevans


git-svn-id: https://svn.freebsd.org/base/stable/11@331114 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f