Bugfix: wait for writes/erases to complete after starting them, instead of before starting them.
Using the wait-before logic would make sense if there was useful time-
consuming work that could be done between the end of one write and the
beginning of the next, but it also requires doing the wait-for-ready before
reading, because a prior write or erase could still be in progress.  Reading
is the far more common case, so adding a whole extra bus transaction to
check for ready before each read would soak up any small gains that might be
had from doing async writes.


git-svn-id: https://svn.freebsd.org/base/head@331132 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f