Forth version of EFI autoresizing
r331321 delegated autoresizing to an efi-autoresizecons command that
currently is expected to be done in forth/lua prior to drawing anything
useful.

Add the Forth version of the lua addition in r331321, hook efi.4th up to be
installed.

efiboot? was written by dteske@; anything outside of that may be blamed on
me.


git-svn-id: https://svn.freebsd.org/base/head@331326 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f