Lock reservations with a dedicated lock in each reservation.  Protect the vmd_free_count with atomics.
This allows us to allocate and free from reservations without the free lock
except where a superpage is allocated from the physical layer, which is
roughly 1/512 of the operations on amd64.

Use the counter api to eliminate cache conention on counters.

Reviewed by:	markj
Tested by:	pho
Sponsored by:	Netflix, Dell/EMC Isilon
Differential Revision:	https://reviews.freebsd.org/D14707


git-svn-id: https://svn.freebsd.org/base/head@331369 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f