MFC 328909: Always give ELF brands a chance to veto a match.
If a brand provides a header_supported hook, check it when trying to
find a brand based on a matching interpreter as well as in the final
loop for the fallback brand. Previously a brand might reject a binary
via a header_supported hook in one of the earlier loops, but still be
chosen by one of these later loops.

Sponsored by:	DARPA / AFRL


git-svn-id: https://svn.freebsd.org/base/stable/11@331458 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f