MFC r327756-r327758
r327756:
Bugfix: on RTC chips with a 32-bit binary counter, after reading the time,
return immediately rather than falling through to the logic that reads
BCD-encoded time.

r327757:
Bugfix: don't lose the am/pm mode flag when setting the time.  Unlike most
RTC chips that have a control register bit for am/pm mode, the DS13xx series
uses one of the high bits in the hour register.  Thus, when setting the time
in am/pm mode, the am/pm mode flag has to be ORed into the hour.

r327758:
Convert a collection of unrelated bitwise flags to a collection of boolean
vars in the softc.  It makes the code more compact and readable, and
actually uses less memory too.


git-svn-id: https://svn.freebsd.org/base/stable/11@331495 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f