MFC r330385:
Flag the first interface on a KTLINK FTDI-based jtag+uart device as being
the jtag port, so that a tty is not created for it.

This is based on information in the PR and from the vendor website.  When
the PR was first opened we had no facility for flagging the jtag ports.  I
stumbled across the still-open PR with the idea of closing it, and noticed
that this wee update was needed.

PR:		175893


git-svn-id: https://svn.freebsd.org/base/stable/11@331509 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f