MFC r331274
Extend the description of ALTQ to call it a system which is a framework in
altq(4) to match altq(9). This makes preserving the history section as the
author of ALTQ easier in the history section, rather than calling it a framework
in the description & a system in the history.
Add a history section to altq(4) and extend the history section in altq(9)


git-svn-id: https://svn.freebsd.org/base/stable/11@331511 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f