MFC 329785: Move DDP PCB state into a helper structure.
This consolidates all of the DDP state in one place.  Also, the code has
now been fixed to ensure that DDP state is only accessed for DDP
connections.  This should not be a functional change but makes it cleaner
and easier to add state for other TOE socket modes in the future.

Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/11@331645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f